finish Datepicker tests
